Fix signed missmatch compairsons for child indices in OPL_Queue_Pop

+++ opl/opl_queue.cpp
@@ -145,13 +145,13 @@ int OPL_Queue_Pop(opl_callback_queue_t *queue,
         child1 = i * 2 + 1;
         child2 = i * 2 + 2;
 
-        if (child1 < queue->num_entries
+        if (child1 < static_cast<int>(queue->num_entries)
          && queue->entries[child1].time < entry->time)
         {
             // Left child is less than entry.
             // Use the minimum of left and right children.
 
-            if (child2 < queue->num_entries
+            if (child2 < static_cast<int>(queue->num_entries)
              && queue->entries[child2].time < queue->entries[child1].time)
             {
                 next_i = child2;
@@ -161,7 +161,7 @@ int OPL_Queue_Pop(opl_callback_queue_t *queue,
                 next_i = child1;
             }
         }
-        else if (child2 < queue->num_entries
+        else if (child2 < static_cast<int>(queue->num_entries)
               && queue->entries[child2].time < entry->time)
         {
             // Right child is less than entry.  Go down the right side.
@@ -205,7 +205,7 @@ void OPL_Queue_AdjustCallbacks(opl_callback_queue_t *queue,
                                uint64_t time, float factor)
 {
     int64_t offset;
-    int i;
+    unsigned int i;
 
     for (i = 0; i < queue->num_entries; ++i)
     {
